+++ beam/beam-core/src/main/java/org/esa/beam/framework/datamodel/CrsGeoCoding.java
@@ -220,7 +220,7 @@ public class CrsGeoCoding extends AbstractGeoCoding {
         }
         try {
             DirectPosition directGeoPos = imageToGeo.transform(new DirectPosition2D(pixelPos), null);
-            geoPos.setLocation(directGeoPos.getOrdinate(1), directGeoPos.getOrdinate(0));
+            geoPos.setLocation((float) directGeoPos.getOrdinate(1), (float) directGeoPos.getOrdinate(0));
         } catch (Exception ignored) {
             geoPos.setInvalid();
         }
@@ -228,7 +228,7 @@ public class CrsGeoCoding extends AbstractGeoCoding {
     }
 
     public final void getPixels(final int x1, final int y1, final int w, final int h,
-                                  final double[] latPixels, final double[] lonPixels) {
+                                  final float[] latPixels, final float[] lonPixels) {
         final DirectPosition2D directPixPos = new DirectPosition2D();
         final DirectPosition directGeoPos = new GeneralDirectPosition(0,0);
         final int x2 = x1 + w;
@@ -240,11 +240,11 @@ public class CrsGeoCoding extends AbstractGeoCoding {
                 try {
                     directPixPos.setLocation(x + 0.5, yp);
                     imageToGeo.transform(directPixPos, directGeoPos);
-                    latPixels[pos] = directGeoPos.getOrdinate(1);
-                    lonPixels[pos] = directGeoPos.getOrdinate(0);
+                    latPixels[pos] = (float) directGeoPos.getOrdinate(1);
+                    lonPixels[pos] = (float) directGeoPos.getOrdinate(0);
                 } catch (Exception ignored) {
-                    latPixels[pos] = Double.NaN;
-                    lonPixels[pos] = Double.NaN;
+                    latPixels[pos] = Float.NaN;
+                    lonPixels[pos] = Float.NaN;
                 }
                 ++pos;
             }
@@ -258,7 +258,7 @@ public class CrsGeoCoding extends AbstractGeoCoding {
         }
         try {
             DirectPosition directPixelPos = geoToImage.transform(new DirectPosition2D(geoPos.getLon(), geoPos.getLat()), null);
-            pixelPos.setLocation(directPixelPos.getOrdinate(0), directPixelPos.getOrdinate(1));
+            pixelPos.setLocation((float) directPixelPos.getOrdinate(0), (float) directPixelPos.getOrdinate(1));
         } catch (Exception ignored) {
             pixelPos.setInvalid();
         }
