+++ beam/beam-visat-rcp/src/main/java/org/esa/beam/visat/actions/ExportEnviGcpFileAction.java
@@ -106,8 +106,8 @@ public class ExportEnviGcpFileAction extends ExecCommand {
                     pixelPos.x = imageX + 0.5f;
                     pixelPos.y = imageY + 0.5f;
                     geoCoding.getGeoPos(pixelPos, geoPos);
-                    final double mapX = geoPos.lon; //longitude
-                    final double mapY = geoPos.lat; //latitude
+                    final float mapX = geoPos.lon; //longitude
+                    final float mapY = geoPos.lat; //latitude
                     writer.write(createLineString(mapX, mapY,
                                                   pixelPos.x + 1,
                                                   // + 1 because ENVI uses a one-based pixel co-ordinate system
@@ -126,7 +126,7 @@ public class ExportEnviGcpFileAction extends ExecCommand {
         return str.concat(_GCP_LINE_SEPARATOR);
     }
 
-    private static String createLineString(final double mapX, final double mapY, final double imageX, final double imageY) {
+    private static String createLineString(final float mapX, final float mapY, final float imageX, final float imageY) {
         return "" + mapX + "\t" + mapY + "\t" + imageX + "\t" + imageY + _GCP_LINE_SEPARATOR;
     }
